Experiment with `constexpr` traits in `enable_if`.
So the `constexpr` traits do not really look better here.  But
they may be better for writing linear code in complex trait
processing, as opposed to complicated recursive code.  Still
to be determined.  However, If I stay with this and if
`std::find_if` and friends are `constexpr`, then the `Mata`
algorithms can be removed, and `std::` forms will just work.
That's a nice bonus.

Otherwise, we need something more like `boost::mpl`, which is
kinda annoying.  Still, `Meta::Container::vector` is very
nice -- it side-steps some issues with using `std::tuple` as
a type-list.